+++ server/src/Handlers/Petrinet/RegisterPetrinet.php
@@ -7,7 +7,7 @@ use Psr\Http\Message\ResponseInterface as Response;
 use Slim\Exception\HttpBadRequestException;
 
 use Cora\Handlers\AbstractHandler;
-use Cora\Services\RegisterPetrinetService;
+use Cora\Service\RegisterPetrinetService;
 use Cora\Views\Factory\ViewFactory;
 use Cora\Views\Factory\PetrinetCreatedViewFactory;
