+++ server/src/Handlers/User/RegisterUser.php
@@ -7,7 +7,7 @@ use Psr\Http\Message\ResponseInterface as Response;
 use Slim\Exception\HttpBadRequestException;
 
 use Cora\Handlers\AbstractHandler;
-use Cora\Services\RegisterUserService;
+use Cora\Service\RegisterUserService;
 use Cora\Views\Factory\ViewFactory;
 use Cora\Views\Factory\UserCreatedViewFactory;
