+++ server/src/Service/RegisterUserService.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Cora\Services;
+namespace Cora\Service;
 
 use Cora\Repositories\UserRepository;
 use Cora\Domain\User\Validation\UniqueUserRule;
